pyproject.toml
[tool.poetry]
name = "datasearch"
version = "0.1.0"
description = ""
authors = ["Bhavin <bhavin200@hotmail.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sentence-transformers = "^3.0.1"
numpy = "1.26.4"
faiss-cpu = "^1.8.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
